Change BasicExample to StringExample
